Fixed PoD and PoY when less than 10
Edge case during the first minutes of the day or first days of the year,
when the number used to derive the percentage is less than 10, so
getting the substring of the first two and third characters won't work,
nor will the crude padding code work. Fixed now, using a spicy while
loop.