feat(sldrclk): expanding update logic
We will treat the watchface as two separate components: slow and fast.
Fast will be updated every 500 ms, while the slow will be updated once
a minute. There is some background math needed to get this to work, but
it should be way nicer on battery economy than needing to redraw (and
recalculate) all the elements twice a second.